Add files via upload
This routines, although still hard-coded, can produce the csv file exactly the same as the one that I (Howard) send out on 16 Nov 2021 evening.  Data on patent has been added.  The root cause of misalignment is not solved, but rectifications were made so that the 'consolidated_data' file in database has correct shape and values.